LayoutTests/inspector/script-profiler/event-type-API-expected.txt

CONSOLE MESSAGE: line 1: Log from API Evaluated Script
Tests for generating a ScriptProfiler.Event with ScriptProfiler.EventType.API type.


== Running test suite: ScriptProfiler.EventType.API
-- Running test case: TrackingSession
ScriptProfiler.trackingStart
ScriptProfiler.trackingUpdate
PASS: Event type should be API.
ScriptProfiler.trackingComplete
